Move transfer_functions and tone_mapping to cms/
Supporting headers fast_math and rational_polynimial look generic enough -> moved to base/
