+++ lib/jxl/base/fast_math-inl.h
@@ -5,17 +5,17 @@
 
 // Fast SIMD math ops (log2, encoder only, cos, erf for splines)
 
-#if defined(LIB_JXL_FAST_MATH_INL_H_) == defined(HWY_TARGET_TOGGLE)
-#ifdef LIB_JXL_FAST_MATH_INL_H_
-#undef LIB_JXL_FAST_MATH_INL_H_
+#if defined(LIB_JXL_BASE_FAST_MATH_INL_H_) == defined(HWY_TARGET_TOGGLE)
+#ifdef LIB_JXL_BASE_FAST_MATH_INL_H_
+#undef LIB_JXL_BASE_FAST_MATH_INL_H_
 #else
-#define LIB_JXL_FAST_MATH_INL_H_
+#define LIB_JXL_BASE_FAST_MATH_INL_H_
 #endif
 
 #include <hwy/highway.h>
 
 #include "lib/jxl/base/common.h"
-#include "lib/jxl/rational_polynomial-inl.h"
+#include "lib/jxl/base/rational_polynomial-inl.h"
 HWY_BEFORE_NAMESPACE();
 namespace jxl {
 namespace HWY_NAMESPACE {
@@ -216,11 +216,11 @@ V CubeRootAndAdd(const V x, const V add) {
 }  // namespace jxl
 HWY_AFTER_NAMESPACE();
 
-#endif  // LIB_JXL_FAST_MATH_INL_H_
+#endif  // LIB_JXL_BASE_FAST_MATH_INL_H_
 
 #if HWY_ONCE
-#ifndef FAST_MATH_ONCE
-#define FAST_MATH_ONCE
+#ifndef LIB_JXL_BASE_FAST_MATH_ONCE
+#define LIB_JXL_BASE_FAST_MATH_ONCE
 
 namespace jxl {
 inline float FastLog2f(float f) { return HWY_STATIC_DISPATCH(FastLog2f)(f); }
@@ -232,5 +232,5 @@ inline float FastCosf(float f) { return HWY_STATIC_DISPATCH(FastCosf)(f); }
 inline float FastErff(float f) { return HWY_STATIC_DISPATCH(FastErff)(f); }
 }  // namespace jxl
 
-#endif  // FAST_MATH_ONCE
+#endif  // LIB_JXL_BASE_FAST_MATH_ONCE
 #endif  // HWY_ONCE
