+++ lib/jxl/butteraugli/butteraugli.cc
@@ -37,10 +37,10 @@
 #define HWY_TARGET_INCLUDE "lib/jxl/butteraugli/butteraugli.cc"
 #include <hwy/foreach_target.h>
 
+#include "lib/jxl/base/fast_math-inl.h"
 #include "lib/jxl/base/printf_macros.h"
 #include "lib/jxl/base/status.h"
 #include "lib/jxl/convolve.h"
-#include "lib/jxl/fast_math-inl.h"
 #include "lib/jxl/gauss_blur.h"
 #include "lib/jxl/image_ops.h"
