+++ lib/jxl/enc_ans.cc
@@ -19,13 +19,13 @@
 
 #include "lib/jxl/ans_common.h"
 #include "lib/jxl/base/bits.h"
+#include "lib/jxl/base/fast_math-inl.h"
 #include "lib/jxl/dec_ans.h"
 #include "lib/jxl/enc_aux_out.h"
 #include "lib/jxl/enc_cluster.h"
 #include "lib/jxl/enc_context_map.h"
 #include "lib/jxl/enc_fields.h"
 #include "lib/jxl/enc_huffman.h"
-#include "lib/jxl/fast_math-inl.h"
 #include "lib/jxl/fields.h"
 
 namespace jxl {
