+++ lib/jxl/enc_cluster.cc
@@ -20,7 +20,7 @@
 #include <hwy/highway.h>
 
 #include "lib/jxl/ac_context.h"
-#include "lib/jxl/fast_math-inl.h"
+#include "lib/jxl/base/fast_math-inl.h"
 HWY_BEFORE_NAMESPACE();
 namespace jxl {
 namespace HWY_NAMESPACE {
