+++ lib/jxl/fast_math_test.cc
@@ -8,11 +8,11 @@
 #include <hwy/foreach_target.h>
 
 #include "lib/jxl/base/random.h"
+#include "lib/jxl/cms/transfer_functions-inl.h"
 #include "lib/jxl/dec_xyb-inl.h"
 #include "lib/jxl/enc_xyb.h"
 #include "lib/jxl/jxl_cms.h"
 #include "lib/jxl/testing.h"
-#include "lib/jxl/transfer_functions-inl.h"  // everything
 
 // Test utils
 #include <hwy/highway.h>
