+++ lib/jxl/jxl_cms.cc
@@ -37,11 +37,11 @@
 #include "lib/jxl/base/printf_macros.h"
 #include "lib/jxl/base/span.h"
 #include "lib/jxl/base/status.h"
+#include "lib/jxl/cms/transfer_functions-inl.h"
 #include "lib/jxl/color_encoding_internal.h"
 #include "lib/jxl/color_management.h"
 #include "lib/jxl/field_encodings.h"
 #include "lib/jxl/image.h"
-#include "lib/jxl/transfer_functions-inl.h"  // TF_HLG TF_PQ
 #if JPEGXL_ENABLE_SKCMS
 #include "lib/jxl/jxl_skcms.h"
 #else  // JPEGXL_ENABLE_SKCMS
