+++ lib/jxl/rational_polynomial_test.cc
@@ -13,8 +13,8 @@
 #include <hwy/tests/hwy_gtest.h>
 
 #include "lib/jxl/base/common.h"
+#include "lib/jxl/base/rational_polynomial-inl.h"
 #include "lib/jxl/base/status.h"
-#include "lib/jxl/rational_polynomial-inl.h"
 #include "lib/jxl/testing.h"
 
 HWY_BEFORE_NAMESPACE();
