+++ lib/jxl/render_pipeline/stage_from_linear.cc
@@ -10,10 +10,10 @@
 #include <hwy/foreach_target.h>
 #include <hwy/highway.h>
 
+#include "lib/jxl/cms/tone_mapping-inl.h"
+#include "lib/jxl/cms/transfer_functions-inl.h"
 #include "lib/jxl/common.h"  // JXL_HIGH_PRECISION
-#include "lib/jxl/dec_tone_mapping-inl.h"
 #include "lib/jxl/sanitizers.h"
-#include "lib/jxl/transfer_functions-inl.h"  // everything but TF_HLG
 
 HWY_BEFORE_NAMESPACE();
 namespace jxl {
