+++ lib/jxl/render_pipeline/stage_to_linear.cc
@@ -10,9 +10,9 @@
 #include <hwy/foreach_target.h>
 #include <hwy/highway.h>
 
-#include "lib/jxl/dec_tone_mapping-inl.h"
+#include "lib/jxl/cms/tone_mapping-inl.h"
+#include "lib/jxl/cms/transfer_functions-inl.h"
 #include "lib/jxl/sanitizers.h"
-#include "lib/jxl/transfer_functions-inl.h"  //everything but FastLinearToSRGB
 
 HWY_BEFORE_NAMESPACE();
 namespace jxl {
