+++ lib/jxl/tf_gbench.cc
@@ -11,7 +11,7 @@
 #include <hwy/foreach_target.h>
 #include <hwy/highway.h>
 
-#include "lib/jxl/transfer_functions-inl.h"  // TF_PQ TF_SRGB FastLinearToSRGB
+#include "lib/jxl/cms/transfer_functions-inl.h"
 
 HWY_BEFORE_NAMESPACE();
 namespace jxl {
