+++ lib/jxl_lists.cmake
@@ -16,6 +16,7 @@ set(JPEGXL_INTERNAL_BASE_SOURCES
   jxl/base/compiler_specific.h
   jxl/base/data_parallel.cc
   jxl/base/data_parallel.h
+  jxl/base/fast_math-inl.h
   jxl/base/float.h
   jxl/base/iaca.h
   jxl/base/matrix_ops.h
@@ -26,6 +27,7 @@ set(JPEGXL_INTERNAL_BASE_SOURCES
   jxl/base/printf_macros.h
   jxl/base/random.cc
   jxl/base/random.h
+  jxl/base/rational_polynomial-inl.h
   jxl/base/sanitizer_definitions.h
   jxl/base/scope_guard.h
   jxl/base/span.h
@@ -123,6 +125,8 @@ set(JPEGXL_INTERNAL_DEC_SOURCES
   jxl/chroma_from_luma.cc
   jxl/chroma_from_luma.h
   jxl/cms/opsin_params.h
+  jxl/cms/tone_mapping-inl.h
+  jxl/cms/transfer_functions-inl.h
   jxl/codec_in_out.h
   jxl/coeff_order.cc
   jxl/coeff_order.h
@@ -170,7 +174,6 @@ set(JPEGXL_INTERNAL_DEC_SOURCES
   jxl/dec_noise.h
   jxl/dec_patch_dictionary.cc
   jxl/dec_patch_dictionary.h
-  jxl/dec_tone_mapping-inl.h
   jxl/dec_transforms-inl.h
   jxl/dec_xyb-inl.h
   jxl/dec_xyb.cc
@@ -190,7 +193,6 @@ set(JPEGXL_INTERNAL_DEC_SOURCES
   jxl/fast_dct32-inl.h
   jxl/fast_dct64-inl.h
   jxl/fast_dct8-inl.h
-  jxl/fast_math-inl.h
   jxl/field_encodings.h
   jxl/fields.cc
   jxl/fields.h
@@ -251,7 +253,6 @@ set(JPEGXL_INTERNAL_DEC_SOURCES
   jxl/quantizer-inl.h
   jxl/quantizer.cc
   jxl/quantizer.h
-  jxl/rational_polynomial-inl.h
   jxl/render_pipeline/low_memory_render_pipeline.cc
   jxl/render_pipeline/low_memory_render_pipeline.h
   jxl/render_pipeline/render_pipeline.cc
@@ -297,7 +298,6 @@ set(JPEGXL_INTERNAL_DEC_SOURCES
   jxl/splines.h
   jxl/toc.cc
   jxl/toc.h
-  jxl/transfer_functions-inl.h
   jxl/transpose-inl.h
   jxl/xorshift128plus-inl.h
 )
